Moves bitplane collection and shifter out of Chipset.[h/c]pp.
